Move QueryTermSimple to separate h and cpp files.
