+++ searchcore/src/vespa/searchcore/proton/documentmetastore/lid_allocator.cpp
@@ -1,7 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "lid_allocator.h"
-#include <vespa/searchlib/query/queryterm.h>
 #include <vespa/searchlib/common/bitvectoriterator.h>
 #include <mutex>
