+++ searchlib/src/vespa/searchlib/attribute/imported_search_context.cpp
@@ -5,6 +5,7 @@
 #include "imported_attribute_vector.h"
 #include "reference_attribute.h"
 #include <vespa/searchlib/common/bitvectoriterator.h>
+#include <vespa/searchlib/query/queryterm.h>
 #include <vespa/searchlib/queryeval/emptysearch.h>
 #include "attributeiterators.hpp"
