+++ searchlib/src/vespa/searchlib/attribute/multistringpostattribute.hpp
@@ -6,7 +6,7 @@
 #include "multistringpostattribute.h"
 #include "multistringattribute.hpp"
 #include <vespa/fastlib/io/bufferedfile.h>
-#include <vespa/searchlib/query/queryterm.h>
+#include <vespa/searchlib/query/query_term_simple.h>
 
 namespace search {
