+++ searchlib/src/vespa/searchlib/attribute/singlenumericenumattribute.hpp
@@ -9,7 +9,7 @@
 #include "primitivereader.h"
 #include "attributeiterators.hpp"
 #include <vespa/searchlib/queryeval/emptysearch.h>
-#include <vespa/searchlib/query/queryterm.h>
+#include <vespa/searchlib/query/query_term_simple.h>
 #include <vespa/searchlib/util/fileutil.hpp>
 
 namespace search {
