+++ searchlib/src/vespa/searchlib/query/query_term_decoder.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "queryterm.h"
+#include "query_term_simple.h"
 #include <vespa/vespalib/stllike/string.h>
 
 namespace search {
